Version mejorada y ordenada del indexer
