+++ banner.txt
@@ -1,0 +1,17 @@
+
+ ,---,                                                                       
+,`--.' |                    ,---,                                              
+|   :  :       ,---,      ,---.'|                                      __  ,-. 
+:   |  '   ,-+-. /  |     |   | :             ,--,  ,--,             ,' ,'/ /| 
+|   :  |  ,--.'|'   |     |   | |    ,---.    |'. \/ .`|     ,---.   '  | |' | 
+'   '  ; |   |  ,"' |   ,--.__| |   /     \   '  \/  / ;    /     \  |  |   ,' 
+|   |  | |   | /  | |  /   ,'   |  /    /  |   \  \.' /    /    /  | '  :  /   
+'   :  ; |   | |  | | .   '  /  | .    ' / |    \  ;  ;   .    ' / | |  | '    
+|   |  ' |   | |  |/  '   ; |:  | '   ;   /|   / \  \  \  '   ;   /| ;  : |    
+'   :  | |   | |--'   |   | '/  ' '   |  / | ./__;   ;  \ '   |  / | |  , ;    
+;   |.'  |   |/       |   :    :| |   :    | |   :/\  \ ; |   :    |  ---'     
+'---'    '---'         \   \  /    \   \  /  `---'  `--`   \   \  /            
+                        `----'      `----'                  `----'             
+                                                                               
+ZincSearch Indexer 1.0 
+Powered by Elvin Nájera
